requirements.txt
flask==3.1.0
pywebview==5.3.2
requests==2.32.3
cryptography==44.0.0
pywin32==308
lxml==5.3.1
openpyxl==3.1.5
pyinstaller==6.12.0
